Added ChangePassword and DeleteUser

+++ controllers/userControllers.js
@@ -5,6 +5,7 @@ const Event = require("../models/Event");
 const EventRegistration = require("../models/EventRegistration");
 const sendMail = require("../utils/sendEmail");
 const generateToken = require("../utils/generateToken");
+const axios = require('axios');
 const asyncHandler = require("express-async-handler");
 
 const {
@@ -229,11 +230,81 @@ const unregisterUserForEvent = async (req, res) => {
   }
 };
 
+// Written by PARAS NAGPAL :( 
 const deleteUser = async (req, res) => {
   try {
     const userId = req.params.userId;
-    const user = await User.findOneAndDelete({ _id: userId });
-    res.status(200).json(user);
+
+    const authorizationHeader = req.headers['authorization'];
+
+    if (!authorizationHeader) {
+      return res.status(401).json({ error: 'Unauthorized - Missing Authorization header' });
+    }
+  
+    const token = authorizationHeader.split(' ')[1];
+
+    try {
+      const headers = {
+        Authorization: `Bearer ${token}`,
+      };
+      const response = await axios.get('http://localhost:3000/api/v1/events', { headers });
+      const allEvents = [];
+
+      for (const event of response.data) {
+        allEvents.push(event.event_id);
+      }
+ 
+      for (const eventId of allEvents) {
+        try {
+          const eventRegistration = await EventRegistration.findOne({ eventId });
+          
+          if (eventRegistration) {
+            const users = eventRegistration.users;
+            
+            for (const user of users) {
+              if (user.userId === userId) {
+                await EventRegistration.updateOne(
+                  { eventId: eventId },
+                  { $pull: { users: { userId: userId } } }
+                );
+                
+                console.log('User deleted from event registration');
+              }
+            }
+          }
+        } catch (error) {
+          console.error('Error updating event registration:', error);
+        }
+      }
+      
+     
+     const friends_userID = [] ;
+     for(const friend of user.friends){
+        friends_userID.push(friend.userId);
+     }
+     for(const id of friends_userID){
+        if(id == userId){
+            await User.updateOne(
+              { _id: userId },
+              { $pull: { friends: { userId: userId } } },
+              (err, result) => {
+                if (err) {
+                  console.error('Error updating user:', err);
+                } else {
+                  console.log('User deleted from friends:', result);
+                }
+              })
+        }
+     }
+    
+     await UserInterest.findOneAndDelete({_id : userId});
+     await UserRegistration.findOneAndDelete({_id : userId});
+     const user = await User.findOneAndDelete({ _id: userId });
+     res.status(200).json(user);
+    } catch (error) {
+      console.error('Error fetching data:', error.message);
+      res.status(500).json({ error: 'Internal Server Error' });
+    }
   } catch (error) {
     console.error("Error:", error.message);
     return res
@@ -290,7 +361,6 @@ const changePassword = async (req, res) => {
     if (user && (await user.matchPassword(oldPassword))) {
       user.password = newPassword;
       const updatedUser = await user.save();
-
       res
         .status(200)
         .json({ message: "Password updated successfully", user: updatedUser });
@@ -299,7 +369,6 @@ const changePassword = async (req, res) => {
     }
   } catch (error) {
     console.error("Error:", error.message);
-
     res
       .status(500)
       .json({ error: "Internal Server Error", details: error.message });
